add taxss

+++ app/Http/Controllers/API/LengthclassController.php
@@ -46,7 +46,7 @@ class LengthclassController extends Controller
        $this->validate($request, [
             'title' => 'required|max:32', 
             'code' => 'required|string|max:4',
-            'value' => 'integer'
+            'value' => 'numeric'
             ]);  
           
        $created= Lengthclass::create([
@@ -89,7 +89,7 @@ class LengthclassController extends Controller
         $this->validate($request, [
             'title' => 'required|max:32', 
             'code' => 'required|string|max:4',
-            'value' => 'integer',
+            'value' => 'numeric',
             ]);
